+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.3]
+
+Features:
+
+- Now uses the dicom-anonymizer library to make a local copy of each scan with
+  identifying metadata fields stripped out or replaced with appropriate values.
+  The default anonymisation behaviour can be configured with an AllowFields
+  parameter.
+
 ## [1.1.2]
 
 Bug fixes: